Finished both methods, program works
Implemented both methods and tested simple input cases that are known to
work, though the word ladders aren't always the shortest. However, the
guide said we didn't have to find the shortest one, just that there IS
one. Also included test1.txt file as input parameter for the test cases.
Need to do more testing and commenting, and also start the deliverables.